requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
httpx
PyJWT
requests
psycopg[binary]
